Removed csnContext.ContextData
Removing the ContextData class is necessary to allow the development version of
csnGUI to use csnContext. This is important for refactoring: I want
to work on a new version of the GUI (based on the dev branch)
that works with the main branch, because in the end this will allow us to
replace the old gui with the new gui in the main branch, without changing the csnake core.

Also, since there is no need to have a separate ContextData class, it is better to simplify the code and remove it.
Apparantly, csnContext.ContextData.Equal was used to be compare the old context
data to the new context data in the GUI. This comparison is now achieved
with the new csnContext.Dump function (csnGUI uses this function to compare
the new Dump output to the previous output).

Finally, I added a cmakeVersion field in the context. The next GUI version will use
this field to remember which version of CMake was used to configure your code.