Fix third party folder index search on Windows with inconsistent letter casing in paths (#104).
On Windows, two paths that only differ in their letter casing are equally valid and refer to the same file or directory. Depending on the source of the paths, one casing or the other may be obtained. This breaks the search for the third party folder entry, to which a given project belongs, since it compares the paths case-sensitive.

The solution is not as simple as switching to a case-insensitive comparison because on Unix-like operating systems two paths that are equal apart from they casing may refer to different directories. If they do may even depend on the mount options, so there is no way to check equality on a string level.

Instead, use "os.stat" to compare the actual identity of the files/directories. The limitation, that the comparison only works for files/directories that actually exist, is inherent of the problem type and not a real limitation in this context (third party source folders have to exist already).