Use os.path.samefile (Unix) or a normalized string comparison (Windows) in "IsSameFileOrDirectory" (#104).
The "os.stat" based comparison doesn't work 100% reliably on Windows because the returned object doesn't contain an "inode" id, but instead only metadata that could be the same for different directories in some cases. So fall back to string comparisons on Windows and use "os.path.samefile", wherever available (on Unix for Python < 3 and on all systems for Python >= 3).